feat(GUI): Updated the easeInOutSine easing meehod to use std::cos instead of std::cosf.

+++ gui/include/Maths.hpp
@@ -22,7 +22,7 @@ namespace Maths {
 
     constexpr float easeInOutSine(float value)
     {
-        return -(std::cosf(std::numbers::pi_v<float> * value) - 1) / 2;
+        return -(std::cos(std::numbers::pi_v<float> * value) - 1) / 2;
     }
 
 } // namespace Maths
